mysite/requirement.txt
asgiref==3.3.4
Django==3.2
django-braces==1.14.0
django-model-utils==4.1.1
django-star-ratings==0.9.1
django-widget-tweaks==1.4.8
python-decouple==3.4
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
swapper==1.1.2.post1
